+++ src/components/Test/TestResolver.js
@@ -84,8 +84,25 @@ const TestResolver = {
         throw new GraphQLError(error.message);
       }
     },
-    deleteTest: async (_, { testId }, { token, dataSources: { testAPI } }) => {
+    deleteTest: async (
+      _,
+      { testId },
+      { token, dataSources: { testAPI, contentAPI, bundleContentAPI } }
+    ) => {
       try {
+        const contentExists = await contentAPI.mediaContentExists({
+          media: testId,
+        });
+        const bundleContentExists =
+          await bundleContentAPI.mediaBundleContentExists({
+            media: testId,
+          });
+        if (contentExists || bundleContentExists)
+          throw new GraphQLError(
+            `${
+              contentExists ? "Content" : "Course content"
+            } is using this test, Please remove from that before deleting this.`
+          );
         const payload = await testAPI.deleteTest({ testId });
         return {
           code: "200",
